// src/pages/transporters/truck-master/trucks-list/trucks.types.ts
export interface Truck {
  profileId: number;
  truckNumber: string;
  registrationNumber: string;
  truckType: string;
  make: string;
  model: string;
  manufactureYear: number;
  capacity: number;
  capacityUnit: string;
  fuelType: string;
  ownershipType: string;
  actionPerformedBy: number;
}